+++ src/app/(main)/admin/_components/index.ts
@@ -1,0 +1,1 @@
+export { AdminSidebar } from './Sidebar'
