+++ src/app/(main)/admin/layout.tsx
@@ -6,7 +6,7 @@ import { useRouter } from 'next/navigation'
 
 import { useMyInfoStore } from '@/store/myInfo'
 
-import { AdminSidebar } from './_components/AdminSidebar'
+import { AdminSidebar } from './_components'
 
 const AdminLayout = ({ children }: { children: React.ReactNode }) => {
   const { role } = useMyInfoStore((state) => state.getMyInfo())
