+++ src/app/(main)/admin/member/_components/approve-table/index.ts
@@ -1,0 +1,1 @@
+export { ApproveMemberTable } from './Table'
