--- src/app/(main)/admin/member/_components/approve/index.ts
@@ -1,2 +1,0 @@
-export { ApproveMemberForm } from './Form'
-export { ApproveMemberTable } from './Table'
