+++ src/app/(main)/admin/member/_components/expel-table/index.ts
@@ -1,0 +1,1 @@
+export { ExpelMemberTable } from './Table'
