--- src/app/(main)/admin/member/_components/expel/index.ts
@@ -1,3 +1,0 @@
-export { ExpelMemberDialog } from './Dialog'
-export { ExpelMemberDialogForm } from './DialogForm'
-export { ExpelMemberTable } from './Table'
