+++ src/app/(main)/admin/member/_components/index.ts
@@ -1,3 +1,3 @@
-export * from './approve'
-export * from './change-role'
-export * from './expel'
+export * from './approve-table'
+export * from './change-role-table'
+export * from './expel-table'
