+++ src/app/(main)/admin/member/page.tsx
@@ -1,6 +1,5 @@
+import { SectionWithTitle } from '@/components/feature'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/tabs'
-
-import { SectionWithTitle } from '~admin/_components/SectionWithTitle'
 
 import {
   ApproveMemberTable,
