+++ src/app/(main)/admin/semester/_components/activity-accordion/index.ts
@@ -1,0 +1,1 @@
+export { ActivityAccordion } from './ActivityAccordion'
