+++ src/app/(main)/admin/semester/_components/semester-list/index.ts
@@ -1,0 +1,1 @@
+export { SemesterList } from './SemesterList'
