+++ src/app/(main)/admin/semester/_components/semester-skeleton/index.ts
@@ -1,0 +1,1 @@
+export { SemesterSkeleton } from './SemesterSkeleton'
