+++ src/components/feature/index.ts
@@ -1,0 +1,2 @@
+export * from './member-table'
+export { SectionWithTitle } from './SectionWithTitle'
