+++ src/components/feature/member-table/index.ts
@@ -1,0 +1,2 @@
+export { MemberTable } from './MemberTable'
+export { SkeletonTable } from './SkeletonTable'
